+++ 950-ng-core/src/main/java/io/harness/ng/core/envGroup/beans/EnvironmentGroupConfig.java
@@ -1,0 +1,29 @@
+package io.harness.ng.core.envGroup.beans;
+
+import static io.harness.annotations.dev.HarnessTeam.PIPELINE;
+
+import io.harness.annotations.dev.OwnedBy;
+import io.harness.data.validator.EntityIdentifier;
+import io.harness.data.validator.EntityName;
+import io.harness.ng.core.common.beans.NGTag;
+
+import java.util.List;
+import lombok.Builder;
+import lombok.Data;
+
+@OwnedBy(PIPELINE)
+@Data
+@Builder
+public class BasicEnvironmentGroup {
+  @EntityName String name;
+  @EntityIdentifier String identifier;
+
+  String orgIdentifier;
+  String projectIdentifier;
+
+  String description;
+  String color;
+  List<NGTag> tags;
+
+  private List<String> envIdentifiers;
+}
